requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email-validator]==2.9.2
python-multipart==0.0.12

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
